app-streamlit/pyproject.toml
[project]
name = "app-streamlit"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.0",
    "colorama>=0.4.6",
    "cryptography>=45.0.2",
    "orjson>=3.10.18",
    "pydantic>=2.11.5",
    "pylint>=3.3.7",
    "python-dotenv>=1.1.0",
    "streamlit>=1.45.1",
    "streamlit-cookies-controller>=0.0.4",
    "websocket-client>=1.8.0",
]
